requirements.txt
pytelebotapi
python-dotenv
supabase
deep-translator
googletrans==4.0.0-rc1
asyncio
